requirements.txt
pandas>=1.5.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
scipy>=1.7.0

# AI and Narrative Analysis
google-generativeai>=0.3.0

# Optional API dependencies
Flask>=2.0.0